+++ library/forms/validators/LengthValidator.php
@@ -1,7 +1,7 @@
 <?php
-require_once(__DIR__.'/ZB_AbstractValidator.php');
+namespace Zerobase\Forms\Validators;
 
-class ZB_LengthValidator extends ZB_AbstractValidator
+class LengthValidator extends AbstractValidator
 {
     protected function getDefaultMessages()
     {
