+++ library/forms/validators/RangeValidator.php
@@ -1,7 +1,7 @@
 <?php
-require_once(__DIR__.'/ZB_AbstractValidator.php');
+namespace Zerobase\Forms\Validators;
 
-class ZB_RangeValidator extends ZB_AbstractValidator
+class RangeValidator extends AbstractValidator
 {
     protected function getDefaultMessages()
     {
